API

Missing docstring.

Missing docstring for JustPIC._2D.setelement!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for JustPIC._3D.element. Check Documenter's build log for details.

Missing docstring.

Missing docstring for JustPIC._3D.setelement!. Check Documenter's build log for details.

JustPIC._3D.@idx — Macro

@idx(args...)

Make a linear range from 1 to args[i], with i ∈ [1, ..., n]

JustPIC._2D.@idx — Macro

@idx(args...)

Make a linear range from 1 to args[i], with i ∈ [1, ..., n]

Missing docstring.

Missing docstring for JustPIC._2D.element. Check Documenter's build log for details.

JustPIC._3D.lerp — Function

lerp(v, t::NTuple{nD,T}) where {nD,T}

Linearly interpolates the value v between the elements of the tuple t. This function is specialized for tuples of length nD.

Arguments

v: The value to be interpolated.

t: The tuple of values to interpolate between.

JustPIC._2D.lerp — Function

lerp(v, t::NTuple{nD,T}) where {nD,T}

Linearly interpolates the value v between the elements of the tuple t. This function is specialized for tuples of length nD.

Arguments

v: The value to be interpolated.

t: The tuple of values to interpolate between.